requirements.txt
absl-py==1.0.0
altair==4.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
bleach==3.3.0
blinker==1.4
boto3>=1.26.0
botocore>=1.26.0
cachetools==4.1.0
certifi==2022.12.7
chardet==3.0.4
click==7.1.1
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
gast==0.2.2
google-auth>=1.25.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2>=3.0
jmespath==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==6.5.1
nbformat==5.0.6
notebook==6.4.12
numpy==1.22.0
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.3
pandas==1.2.0
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.3.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.18.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.3.0
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pywinpty==0.5.7
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa>=4.1
s3transfer>=0.3.3
scipy>=1.6.0
Send2Trash==1.5.0
six==1.14.0
streamlit==0.57.3
tensorboard==2.1.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.11
validators==0.14.3
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
